requirements.txt
invoke
pystache
pyyaml